move GossipNode interface and related types to gossipNode.go
